Plain Action Refactor (#42)
This PR changes the overall strategy for the Zarf action. We want to keep maximum portability for Zarf using in CI so features will be rolled into the Zarf CLI vs a complex Github action. The primary goal of this action will be to make the adoption of Zarf in a workflow frictionless, lightweight and simple. This will be achieved by focusing on simply downloading the Zarf release artifacts as the main purpose of the action.