Update readme and zarf version in test workflow (#44)
Closes https://github.com/defenseunicorns/setup-zarf/issues/43

Signed-off-by: Lucas Rodriguez <lucas.rodriguez@defenseunicorns.com>

- Updates zarf version in readme examples
- Updates example of using setup-zarf action in readme to use pinned
version instead of `main`
- Updates test workflow to use zarf version `v0.24.3`
- Removes event types from `pull_request` event trigger in workflow.
This seemed unnecessary, but I'm okay with adding it back if we want to
keep it.

The scheduled workflow runs were automatically disabled due to 60+ days
of repo inactivity, so they had to be re-enabled. See the
[docs](https://docs.github.com/en/actions/managing-workflow-runs/disabling-and-enabling-a-workflow)
for more information. It would probably be good to get some sort of
automation setup to update the workflow with latest versions of zarf
when they're released for dogfooding and to keep the repo active.
Created an
[issue](https://github.com/defenseunicorns/setup-zarf/issues/45) for
this.